+++ QuickFIXn/NullLogFactory.cs
@@ -1,0 +1,12 @@
+﻿namespace QuickFix
+{
+    public class NullLogFactory : ILogFactory
+    {
+        public NullLogFactory() { }
+
+        public ILog Create(SessionID _x)
+        {
+            return new NullLog();
+        }
+    }
+}