+++ QuickFIXn/Transport/StreamFactory.cs
@@ -15,6 +15,52 @@ namespace QuickFix.Transport
     /// </summary>
     public static class StreamFactory
     {
+        private static Socket CreateTunnelThruProxy(string destIP, int destPort)
+        {
+            string destUriWithPort = $"{destIP}:{destPort}";
+            UriBuilder uriBuilder = new UriBuilder(destUriWithPort);
+            Uri destUri = uriBuilder.Uri;
+            IWebProxy webProxy = WebRequest.GetSystemWebProxy();
+
+            try
+            {
+                if (webProxy.IsBypassed(destUri))
+                    return null;
+            }
+            catch (PlatformNotSupportedException)
+            {
+                // .NET Core doesn't support IWebProxy.IsBypassed
+                // (because .NET Core doesn't have access to Windows-specific services, of course)
+                return null;
+            }
+
+            Uri proxyUri = webProxy.GetProxy(destUri);
+            if (proxyUri == null)
+                return null;
+
+            IPAddress[] proxyEntry = Dns.GetHostAddresses(proxyUri.Host);
+            int iPort = proxyUri.Port;
+            IPAddress address = proxyEntry.First(a => a.AddressFamily == AddressFamily.InterNetwork);
+            IPEndPoint proxyEndPoint = new IPEndPoint(address, iPort);
+            Socket socketThruProxy = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
+            socketThruProxy.Connect(proxyEndPoint);
+
+            string proxyMsg = $"CONNECT {destIP}:{destPort} HTTP/1.1 \n\n";
+            byte[] buffer = Encoding.ASCII.GetBytes(proxyMsg);
+            byte[] buffer12 = new byte[500];
+            socketThruProxy.Send(buffer, buffer.Length, 0);
+            int msg = socketThruProxy.Receive(buffer12, 500, 0);
+            string data;
+            data = Encoding.ASCII.GetString(buffer12);
+            int index = data.IndexOf("200");
+
+            if (index < 0)
+                throw new ApplicationException(
+                    $"Connection failed to {destUriWithPort} through proxy server {proxyUri.ToString()}.");
+
+            return socketThruProxy;
+        }
+
         /// <summary>
         /// Connect to the specified endpoint and return a stream that can be used to communicate with it. (for initiator)
         /// </summary>
@@ -24,15 +70,23 @@ namespace QuickFix.Transport
         /// <returns>an opened and initiated stream which can be read and written to</returns>
         public static Stream CreateClientStream(IPEndPoint endpoint, SocketSettings settings, ILog logger)
         {
-            var socket = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
-            socket.NoDelay = settings.SocketNodelay;
-            if (settings.SocketReceiveBufferSize.HasValue)
+            // If system has configured a proxy for this config, use it.
+            Socket socket = CreateTunnelThruProxy(endpoint.Address.ToString(), endpoint.Port);
+
+            // No proxy.  Set up a regular socket.
+            if (socket == null)
             {
-                socket.ReceiveBufferSize = settings.SocketReceiveBufferSize.Value;
-            }
-            if (settings.SocketSendBufferSize.HasValue)
-            {
-                socket.SendBufferSize = settings.SocketSendBufferSize.Value;
+                socket = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
+                socket.NoDelay = settings.SocketNodelay;
+                if (settings.SocketReceiveBufferSize.HasValue)
+                {
+                    socket.ReceiveBufferSize = settings.SocketReceiveBufferSize.Value;
+                }
+                if (settings.SocketSendBufferSize.HasValue)
+                {
+                    socket.SendBufferSize = settings.SocketSendBufferSize.Value;
+                }
+                socket.Connect(endpoint);
             }
             if (settings.SocketReceiveTimeout.HasValue)
             {
@@ -43,6 +97,7 @@ namespace QuickFix.Transport
                 socket.SendTimeout = settings.SocketSendTimeout.Value;
             }
             socket.Connect(endpoint);
+
             Stream stream = new NetworkStream(socket, true);
 
             if (settings.UseSSL)
@@ -92,9 +147,9 @@ namespace QuickFix.Transport
         {
             X509Certificate2 certificate;
 
-            // TODO: Change to _certificateCache to ConcurrentDictorionary once we start targeting .NET 4
-            // Then remove this lock and use GetOrAdd function of concurrent dictionary
-            //  certificate = _certificateCache.GetOrAdd(name, (key) => LoadCertificateInner(name, password));
+            // TODO: Change _certificateCache's type to ConcurrentDictionary once we start targeting .NET 4,
+            //   then remove this lock and use GetOrAdd function of concurrent dictionary
+            //   e.g.: certificate = _certificateCache.GetOrAdd(name, (key) => LoadCertificateInner(name, password));
             lock (_certificateCache)
             {
                 if (_certificateCache.TryGetValue(name, out certificate))
@@ -142,9 +197,12 @@ namespace QuickFix.Transport
         /// <returns></returns>
         private static X509Certificate2 GetCertificateFromStore(string certName)
         {
+            return GetCertificateFromStoreHelper(certName, new X509Store(StoreLocation.LocalMachine))
+                ?? GetCertificateFromStoreHelper(certName, new X509Store(StoreLocation.CurrentUser));
+        }
 
-            // Get the certificate store for the current user.
-            X509Store store = new X509Store(StoreLocation.CurrentUser);
+        private static X509Certificate2 GetCertificateFromStoreHelper(string certName, X509Store store)
+        {
             try
             {
                 store.Open(OpenFlags.ReadOnly);
@@ -163,14 +221,12 @@ namespace QuickFix.Transport
                 if (currentCerts.Count == 0)
                     return null;
 
-                // Return the first certificate in the collection, has the right name and is current. 
                 return currentCerts[0];
             }
             finally
             {
                 store.Close();
             }
-
         }
 
 
@@ -197,12 +253,12 @@ namespace QuickFix.Transport
             /// <returns>a ssl enabled stream</returns>
             public Stream CreateClientStreamAndAuthenticate(Stream innerStream)
             {
-                var sslStream = new SslStream(innerStream, false, ValidateServerCertificate, SelectLocalCertificate);
+                SslStream sslStream = new SslStream(innerStream, false, ValidateServerCertificate, SelectLocalCertificate);
 
                 try
                 {
                     // Setup secure SSL Communication
-                    var clientCertificates = GetClientCertificates();
+                    X509CertificateCollection clientCertificates = GetClientCertificates();
                     sslStream.AuthenticateAsClient(socketSettings_.ServerCommonName,
                         clientCertificates,
                         socketSettings_.SslProtocol,
@@ -224,7 +280,7 @@ namespace QuickFix.Transport
             /// <returns>a ssl enabled stream</returns>
             public Stream CreateServerStreamAndAuthenticate(Stream innerStream)
             {
-                var sslStream = new SslStream(innerStream, false, ValidateClientCertificate, SelectLocalCertificate);
+                SslStream sslStream = new SslStream(innerStream, false, ValidateClientCertificate, SelectLocalCertificate);
 
                 try
                 {
@@ -232,7 +288,7 @@ namespace QuickFix.Transport
                         throw new Exception(string.Format("No server certificate specified, the {0} setting must be configured", SessionSettings.SSL_CERTIFICATE));
 
                     // Setup secure SSL Communication
-                    var serverCertificate = StreamFactory.LoadCertificate(socketSettings_.CertificatePath, socketSettings_.CertificatePassword);
+                    X509Certificate2 serverCertificate = StreamFactory.LoadCertificate(socketSettings_.CertificatePath, socketSettings_.CertificatePassword);
                     sslStream.AuthenticateAsServer(serverCertificate,
                         socketSettings_.RequireClientCertificate,
                         socketSettings_.SslProtocol,
@@ -252,7 +308,7 @@ namespace QuickFix.Transport
                 if (!string.IsNullOrEmpty(socketSettings_.CertificatePath))
                 {
                     X509CertificateCollection certificates = new X509Certificate2Collection();
-                    var clientCert = StreamFactory.LoadCertificate(socketSettings_.CertificatePath, socketSettings_.CertificatePassword);
+                    X509Certificate2 clientCert = StreamFactory.LoadCertificate(socketSettings_.CertificatePath, socketSettings_.CertificatePassword);
                     certificates.Add(clientCert);
                     return certificates;
                 }
@@ -360,8 +416,8 @@ namespace QuickFix.Transport
                 {
                     if (extension is X509EnhancedKeyUsageExtension)
                     {
-                        var keyUsage = (X509EnhancedKeyUsageExtension)extension;
-                        foreach (var oid in keyUsage.EnhancedKeyUsages)
+                        X509EnhancedKeyUsageExtension keyUsage = (X509EnhancedKeyUsageExtension)extension;
+                        foreach (System.Security.Cryptography.Oid oid in keyUsage.EnhancedKeyUsages)
                         {
                             if (oid.Value == enhancedKeyOid)
                                 return true;
@@ -401,7 +457,5 @@ namespace QuickFix.Transport
                 return localCertificates[0];
             }
         }
-
-
     }
 }