+++ UnitTests/ParserTest.cs
@@ -13,6 +13,11 @@ namespace UnitTests
         const string incomplete_1   = "8=FIX.4.2";
         const string incomplete_2   = "8=FIX.4.2\x01" + "9=12";
 
+        public byte[] StrToBytes(string str)
+        {
+            return CharEncoding.DefaultEncoding.GetBytes(str);
+        }
+
         [Test]
         public void ExtractLength()
         {
@@ -47,7 +52,8 @@ namespace UnitTests
             const string fixMsg3 = "8=FIX.4.2\x01" + "9=19\x01" + "35=A\x01" + "108=30\x01" + "9710=8\x01" + "10=31\x01";
 
             Parser parser = new Parser();
-            parser.AddToStream(fixMsg1 + fixMsg2 + fixMsg3);
+            byte[] combined = StrToBytes(fixMsg1 + fixMsg2 + fixMsg3);
+            parser.AddToStream(combined, combined.Length);
 
             string readFixMsg1;
             Assert.True(parser.ReadFixMessage(out readFixMsg1));
@@ -68,24 +74,28 @@ namespace UnitTests
             string partFixMsg1 = "8=FIX.4.2\x01" + "9=17\x01" + "35=4\x01" + "36=";
             string partFixMsg2 = "88\x01" + "123=Y\x01" + "10=34\x01";
 
-            Parser parser = new Parser();
-            parser.AddToStream(partFixMsg1);
+            byte[] partBytes1 = CharEncoding.DefaultEncoding.GetBytes(partFixMsg1);
+            byte[] partBytes2 = CharEncoding.DefaultEncoding.GetBytes(partFixMsg2);
 
+            Parser parser = new Parser();
             string readPartFixMsg;
+
+            parser.AddToStream(partBytes1, partBytes1.Length);
             Assert.False(parser.ReadFixMessage(out readPartFixMsg));
 
-            parser.AddToStream(partFixMsg2);
+            parser.AddToStream(partBytes2, partBytes2.Length);
             Assert.True(parser.ReadFixMessage(out readPartFixMsg));
+
             Assert.AreEqual(partFixMsg1 + partFixMsg2, readPartFixMsg);
         }
 
         [Test]
         public void ReadFixMessageWithBadLength()
         {
-            string fixMsg = "8=TEST\x01" + "9=TEST\x01" + "35=TEST\x01" + "49=SS1\x01" + "56=RORE\x01" + "34=3\x01" + "52=20050222-16:45:53\x01" + "10=TEST\x01";
+            byte[] fixMsg = StrToBytes("8=TEST\x01" + "9=TEST\x01" + "35=TEST\x01" + "49=SS1\x01" + "56=RORE\x01" + "34=3\x01" + "52=20050222-16:45:53\x01" + "10=TEST\x01");
 
             Parser parser = new Parser();
-            parser.AddToStream(fixMsg);
+            parser.AddToStream(fixMsg, fixMsg.Length);
 
             string readFixMsg;
             Assert.Throws<QuickFix.MessageParseError>(delegate { parser.ReadFixMessage(out readFixMsg); });
@@ -98,23 +108,28 @@ namespace UnitTests
         public void ReadFixMessageWithNonAscii()
         {
             string[] fixMsgFields1 = { "8=FIX.4.4", "9=19", "35=B", "148=Ole!", "33=0", "10=0" };
-            string fixMsg1 = String.Join("\x01", fixMsgFields1) + "\x01";
+            string fixMsg1 = String.Join(Message.SOH, fixMsgFields1) + Message.SOH;
 
             Assert.AreEqual("é", "\x00E9");
             Assert.AreEqual("é", "\xE9");
 
-            string[] fixMsgFields2 = { "8=FIX.4.4", "9=20", "35=B", "148=Olé!", "33=0", "10=0" };
-            string fixMsg2 = String.Join("\x01", fixMsgFields2) + "\x01";
+            // In 1.8 and earlier, the default encoding was UTF-8, which treated "é" as 2 bytes,
+            // and this message had 9=20, which didn't agree with other implementations.
+            // Now that the default encoding is ISO-8859-1, "é" is one byte,
+            // and 9=19.
+            string[] fixMsgFields2 = { "8=FIX.4.4", "9=19", "35=B", "148=Olé!", "33=0", "10=0" };
+            string fixMsg2 = String.Join(Message.SOH, fixMsgFields2) + Message.SOH;
 
             Parser parser = new Parser();
-            parser.AddToStream(fixMsg1 + fixMsg2);
+            byte[] combined = StrToBytes(fixMsg1 + fixMsg2);
+            parser.AddToStream(combined, combined.Length);
 
             string readFixMsg1;
             Assert.True(parser.ReadFixMessage(out readFixMsg1));
             Assert.AreEqual(fixMsg1, readFixMsg1);
 
             string readFixMsg2;
-            Assert.True(parser.ReadFixMessage(out readFixMsg2));
+            Assert.True(parser.ReadFixMessage(out readFixMsg2), "parser.ReadFixMessage(readFixMsg2) failure");
             Assert.AreEqual(fixMsg2, readFixMsg2);
         }
 
@@ -125,7 +140,8 @@ namespace UnitTests
             string fixMsg1 = String.Join("\x01", fixMsgFields1) + "\x01";
 
             Parser parser = new Parser();
-            parser.AddToStream(fixMsg1);
+            byte[] bytesMsg = StrToBytes(fixMsg1);
+            parser.AddToStream(bytesMsg, bytesMsg.Length);
 
             string readFixMsg1;
             Assert.True(parser.ReadFixMessage(out readFixMsg1));