+++ UnitTests/SessionTest.cs
@@ -228,15 +228,16 @@ namespace UnitTests
             session.Next(msg.ToString());
         }
 
-        public bool SENT_RESEND()
+        public bool SENT_SEQUENCE_RESET()
         {
-            if (!responder.msgLookup.ContainsKey(QuickFix.Fields.MsgType.SEQUENCE_RESET) &&
-                responder.msgLookup[QuickFix.Fields.MsgType.SEQUENCE_RESET].Count > 0)
-                return false;
+            return responder.msgLookup.ContainsKey(QuickFix.Fields.MsgType.SEQUENCE_RESET) &&
+                responder.msgLookup[QuickFix.Fields.MsgType.SEQUENCE_RESET].Count > 0;
+        }
 
-            responder.msgLookup[QuickFix.Fields.MsgType.SEQUENCE_RESET].Dequeue();
-
-            return true;
+        public bool SENT_RESEND_REQUEST()
+        {
+            return responder.msgLookup.ContainsKey(QuickFix.Fields.MsgType.RESEND_REQUEST) &&
+                responder.msgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Count > 0;
         }
 
         public bool RESENT()
@@ -434,7 +435,7 @@ namespace UnitTests
             } //seq 4, next is 5
 
             SendResendRequest(1, 4);
-            Assert.That(SENT_RESEND());
+            Assert.That(SENT_SEQUENCE_RESET());
             Assert.IsFalse(RESENT());
         }
 
@@ -491,7 +492,7 @@ namespace UnitTests
             int count = -1;
             foreach (QuickFix.Message sequenceResestMsg in responder.msgLookup[QuickFix.Fields.MsgType.SEQUENCE_RESET])
             {
-                Assert.AreEqual(sequenceResestMsg.GetField(QuickFix.Fields.Tags.GapFillFlag), "Y");
+                Assert.AreEqual(sequenceResestMsg.GetString(QuickFix.Fields.Tags.GapFillFlag), "Y");
                 Assert.AreEqual(sequenceResestMsg.Header.GetInt(QuickFix.Fields.Tags.MsgSeqNum), gapStarts[++count]);
                 Assert.AreEqual(sequenceResestMsg.GetInt(QuickFix.Fields.Tags.NewSeqNo), gapEnds[count]);
             }
@@ -524,7 +525,7 @@ namespace UnitTests
         public void AssertMsInTag(string msgType, int tag, bool shouldHaveMs)
         {
             QuickFix.Message msg = responder.msgLookup[msgType].Last();
-            string sendingTime = msg.Header.GetField(tag);
+            string sendingTime = msg.Header.GetString(tag);
             Match m = msRegex.Match(sendingTime);
             Assert.That(m.Success == shouldHaveMs);
         }
@@ -532,7 +533,7 @@ namespace UnitTests
         public void AssertMicrosecondsInTag(string msgType, int tag, bool shouldHaveMicrosecond)
         {
             QuickFix.Message msg = responder.msgLookup[msgType].Last();
-            string sendingTime = msg.Header.GetField(tag);
+            string sendingTime = msg.Header.GetString(tag);
             Match m = microsecondRegex.Match(sendingTime);
             Assert.That(m.Success == shouldHaveMicrosecond);
         }
@@ -955,5 +956,21 @@ namespace UnitTests
             Assert.That(DISCONNECTED());
         }
 
+        [Test]
+        public void TestResendRequestMsgSeqNumNotIgnoredWhenNoPersistance()
+        {
+            session.PersistMessages = false;
+
+            Logon();
+
+            SendNOSMessage();
+            SendNOSMessage();
+
+            //The below will trigger a sequence reset
+            SendResendRequest(2, 0);
+
+            SendNOSMessage();
+            Assert.That(!SENT_RESEND_REQUEST());
+        }
     }
 }